Improve re-trigger of tests when test sources change
Explicitely handle tracking the source of each test case.
This helps address the fact that examples in specs (in rspec) are not
actual methods but just code blocks and are therefore not automatically
tracked by the tracker when intercepting Tracepoints.

+++ lib/flutter/parser.rb
@@ -8,9 +8,11 @@ module Flutter
     attr_reader :signatures
 
     def initialize(file)
-      code = File.open(file, "r").read
-      @ast = ::Parser::CurrentRuby.parse(code)
-      @signatures = build_signatures(@ast, nil, false)
+      if File.exist?(file)
+        code = File.open(file, "r").read
+        @ast = ::Parser::CurrentRuby.parse(code)
+        @signatures = build_signatures(@ast, nil, false)
+      end
     end
 
     def build_signatures(ast, parent, in_singleton)
